% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largest_component.R
\name{largest_component}
\alias{largest_component}
\title{find the size of the giant component}
\usage{
largest_component(graph)
}
\arguments{
\item{graph}{graph for which to return the giant component}
}
\value{
integer number of nodes in largest component of graph
}
\description{
find the size of the giant component
}

